+++ NEWS.md
@@ -1,6 +1,7 @@
 Red Panda C++ Version 1.0.8
   - enhancement: auto complete '#undef'
   - enhancement: redesign components for compiler commandline arguments processing
+  - fix: selection calculation error when editing in column mode
 
 Red Panda C++ Version 1.0.7
   - change: use Shift+Enter to break line